remove and exculde stax-api in genny-vertical-rules to solve java.lang.LinkageError
details:
https://github.com/quarkusio/quarkus/issues/9300

This problem is actually due to the fact that javax.xml.namespace.QName
is loaded from the QuarkusClassLoader/RunnerClassLoader (because Quarkus
        by default will try to load the classes from the
        QuarkusClassLoader/RunnerClassLoader first) while it should
rather be loaded from the System ClassLoader in that particular case.

To workaround this problem you can indeed exclude the dependency that
causes the LinkageError if you can exclude all the classes of this
dependency otherwise you can rely on the property
quarkus.class-loading.parent-first to indicate Quarkus that its classes
need to be loaded from the parent ClassLoader first before trying to
load it from the QuarkusClassLoader/RunnerClassLoader.

Example of usage of quarkus.class-loading.parent-first in the
application.properties :

quarkus.class-loading.parent-first-artifacts=stax:stax-api